fix: fixed the problem with xmldoc2md
