+++ SlottyMedia.DatabaseSeeding/SlottyMedia.DatabaseSeeding.csproj
@@ -14,7 +14,7 @@
 
     <ItemGroup>
         <PackageReference Include="Bogus" Version="35.6.0"/>
-        <PackageReference Include="DiceBear.Bogus" Version="2024.8.4.912" />
+        <PackageReference Include="DiceBear.Bogus" Version="2024.8.4.912"/>
         <PackageReference Include="Supabase" Version="1.1.1"/>
         <PackageReference Include="WaffleGenerator.Bogus" Version="4.2.1"/>
     </ItemGroup>
