+++ SlottyMedia/Backend/ViewModel/ProfilePageVmImpl.cs
@@ -7,8 +7,8 @@ namespace SlottyMedia.Backend.ViewModel;
 
 public class ProfilePageVmImpl : IProfilePageVm
 {
-    private readonly IUserService _userService;
     private readonly Logging<MainLayoutVmImpl> _logger = new();
+    private readonly IUserService _userService;
 
 
     public ProfilePageVmImpl(IUserService userService)
@@ -19,7 +19,6 @@ public class ProfilePageVmImpl : IProfilePageVm
 
     public async Task<UserInformationDto?> GetUserInfo(Guid userId)
     {
-
         var userDao = await _userService.GetUserDaoById(userId);
         var amountOfFriends = await _userService.GetCountOfUserFriends(userId);
         var amountOfSpaces = await _userService.GetCountOfUserSpaces(userId);
@@ -49,10 +48,7 @@ public class ProfilePageVmImpl : IProfilePageVm
     public async Task<bool?> UserFollowRelation(Guid userIdToCheck, Guid userIdLoggedIn)
     {
         if (userIdToCheck != userIdLoggedIn)
-        {
             return await _userService.UserFollowRelation(userIdToCheck, userIdLoggedIn);
- 
-        }
         return null;
     }
