+++ SlottyMedia/Components/Pages/Home.razor
@@ -15,6 +15,7 @@
 @code
 {
     private Guid _currentUserId;
+
     protected override async Task OnInitializedAsync()
     {
         var _currentSession = AuthService.GetCurrentSession();
