+++ SlottyMedia/Components/Pages/Profile.razor
@@ -1,4 +1,5 @@
 @page "/profile"
+@using System.Web
 @using SlottyMedia.Backend.Dtos
 @using SlottyMedia.Backend.Services.Interfaces
 @using SlottyMedia.Backend.ViewModel.Interfaces
@@ -11,14 +12,14 @@
     <div class="px-4">
         <Avatar Base64PNG="@UserInformationDto.ProfilePic"></Avatar>
     </div>
-    
+
     <div class="px-4 flex flex-col gap-3">
         <h1 class="text-3xl font-medium w-1/4">
             @UserInformationDto.Username
-        </h1> 
-        
+        </h1>
+
         <p class="text-lg">@UserInformationDto.Description</p>
-        
+
         <div class="flex w-1/2">
             <h3 class="text-sm text-cadet-gray w-full">Friends: @UserInformationDto.FriendsAmount</h3>
             <h3 class="text-sm text-cadet-gray w-full">Spaces: @UserInformationDto.SpacesAmount</h3>
@@ -37,11 +38,11 @@
                 {
                     <p>Follow</p>
                 }
-            </HotButton> 
+            </HotButton>
             @_followsUserRetrieved
-        </div> 
+        </div>
     }
-    
+
 </div>
 
 
@@ -51,20 +52,20 @@
     private Session? _currentSession;
 
     private bool? _followsUserRetrieved;
-    
+
     private bool _invalidIdFlag;
-    
+
     public UserInformationDto UserInformationDto { get; set; } = new()
     {
         Username = "Username is loading..",
         Description = "Description is loading..",
         ProfilePic = null
     };
-    
+
     protected override async void OnAfterRender(bool firstRender)
     {
         var uri = new Uri(NavigationManager.Uri);
-        var queryParams = System.Web.HttpUtility.ParseQueryString(uri.Query);
+        var queryParams = HttpUtility.ParseQueryString(uri.Query);
         _currentSession = AuthService.GetCurrentSession();
 
         // Retrieve query parameters
@@ -98,4 +99,5 @@
         _followsUserRetrieved = !_followsUserRetrieved;
         StateHasChanged();
     }
+
 }
