+++ SlottyMedia/Components/Partial/HotButton.razor
@@ -25,10 +25,10 @@
     /// <summary>An event that is triggered when this button is clicked.</summary>
     [Parameter]
     public EventCallback OnClick { get; set; } = EventCallback.Empty;
-    
+
     async Task _handleClick()
     {
-            await OnClick.InvokeAsync();
+        await OnClick.InvokeAsync();
     }
 
 }
