+++ SlottyMedia/Components/Partial/LikeButton.razor
@@ -20,7 +20,7 @@
     [Parameter]
     public bool? Initial { get; set; }
 
-    
+
     /// <summary>
     ///     An event that is triggered when this button is clicked. The callback's parameter
     ///     is a boolean that represents the button's new state.
