+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.commentrepo.commentrepository.md
@@ -5,11 +5,11 @@ Namespace: SlottyMedia.Database.Repository.CommentRepo
 Repository class for managing comments in the database.
 
 ```csharp
-public class CommentRepository : SlottyMedia.Database.Repository.DatabaseRepository`1[[SlottyMedia.Database.Daos.CommentDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.CommentDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], ICommentRepository
+public class CommentRepository : SlottyMedia.Database.Repository.DatabaseRepository`1[[SlottyMedia.Database.Daos.CommentDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.CommentDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], ICommentRepository
 ```
 
 Inheritance [Object](https://docs.microsoft.com/en-us/dotnet/api/system.object) → [DatabaseRepository&lt;CommentDao&gt;](./slottymedia.database.repository.databaserepository-1.md) → [CommentRepository](./slottymedia.database.repository.commentrepo.commentrepository.md)<br>
-Implements [IDatabaseRepository&lt;CommentDao&gt;](./slottymedia.database.idatabaserepository-1.md), [ICommentRepository](./slottymedia.database.repository.commentrepo.icommentrepository.md)
+Implements [IDatabaseRepository&lt;CommentDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md), [ICommentRepository](./slottymedia.database.repository.commentrepo.icommentrepository.md)
 
 ## Constructors
