+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.commentrepo.icommentrepository.md
@@ -5,7 +5,7 @@ Namespace: SlottyMedia.Database.Repository.CommentRepo
 Interface for the Comment Repository.
 
 ```csharp
-public interface ICommentRepository : SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.CommentDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
+public interface ICommentRepository : SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.CommentDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
 ```
 
-Implements [IDatabaseRepository&lt;CommentDao&gt;](./slottymedia.database.idatabaserepository-1.md)
+Implements [IDatabaseRepository&lt;CommentDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md)
