+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.followeruserrelatiorepo.followeruserrelationrepository.md
@@ -5,11 +5,11 @@ Namespace: SlottyMedia.Database.Repository.FollowerUserRelatioRepo
 Repository class for managing follower-user relations in the database.
 
 ```csharp
-public class FollowerUserRelationRepository : SlottyMedia.Database.Repository.DatabaseRepository`1[[SlottyMedia.Database.Daos.FollowerUserRelationDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.FollowerUserRelationDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], IFollowerUserRelationRepository
+public class FollowerUserRelationRepository : SlottyMedia.Database.Repository.DatabaseRepository`1[[SlottyMedia.Database.Daos.FollowerUserRelationDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.FollowerUserRelationDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]], IFollowerUserRelationRepository
 ```
 
 Inheritance [Object](https://docs.microsoft.com/en-us/dotnet/api/system.object) → [DatabaseRepository&lt;FollowerUserRelationDao&gt;](./slottymedia.database.repository.databaserepository-1.md) → [FollowerUserRelationRepository](./slottymedia.database.repository.followeruserrelatiorepo.followeruserrelationrepository.md)<br>
-Implements [IDatabaseRepository&lt;FollowerUserRelationDao&gt;](./slottymedia.database.idatabaserepository-1.md), [IFollowerUserRelationRepository](./slottymedia.database.repository.followeruserrelatiorepo.ifolloweruserrelationrepository.md)
+Implements [IDatabaseRepository&lt;FollowerUserRelationDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md), [IFollowerUserRelationRepository](./slottymedia.database.repository.followeruserrelatiorepo.ifolloweruserrelationrepository.md)
 
 ## Constructors
