+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.forumrepo.iforumrepository.md
@@ -5,10 +5,10 @@ Namespace: SlottyMedia.Database.Repository.ForumRepo
 Interface for the Forum Repository.
 
 ```csharp
-public interface IForumRepository : SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.ForumDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
+public interface IForumRepository : SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.ForumDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
 ```
 
-Implements [IDatabaseRepository&lt;ForumDao&gt;](./slottymedia.database.idatabaserepository-1.md)
+Implements [IDatabaseRepository&lt;ForumDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md)
 
 ## Methods
