+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.postrepo.ipostrepository.md
@@ -5,10 +5,10 @@ Namespace: SlottyMedia.Database.Repository.PostRepo
 Interface for the Post Repository.
 
 ```csharp
-public interface IPostRepository : SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.PostsDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
+public interface IPostRepository : SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.PostsDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
 ```
 
-Implements [IDatabaseRepository&lt;PostsDao&gt;](./slottymedia.database.idatabaserepository-1.md)
+Implements [IDatabaseRepository&lt;PostsDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md)
 
 ## Methods
