+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.searchrepo.iforumsearchrepository.md
@@ -5,10 +5,10 @@ Namespace: SlottyMedia.Database.Repository.SearchRepo
 Interface for the Forum Search Repository.
 
 ```csharp
-public interface IForumSearchRepository : SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.ForumDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
+public interface IForumSearchRepository : SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.ForumDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
 ```
 
-Implements [IDatabaseRepository&lt;ForumDao&gt;](./slottymedia.database.idatabaserepository-1.md)
+Implements [IDatabaseRepository&lt;ForumDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md)
 
 ## Methods
