+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.searchrepo.iuserseachrepository.md
@@ -5,10 +5,10 @@ Namespace: SlottyMedia.Database.Repository.SearchRepo
 Interface for the User Search Repository.
 
 ```csharp
-public interface IUserSeachRepository : SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.UserDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
+public interface IUserSeachRepository : SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.UserDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
 ```
 
-Implements [IDatabaseRepository&lt;UserDao&gt;](./slottymedia.database.idatabaserepository-1.md)
+Implements [IDatabaseRepository&lt;UserDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md)
 
 ## Methods
