+++ slottymedia-docs/docs/code/SlottyMedia.Database/slottymedia.database.repository.userrepo.iuserrepository.md
@@ -5,10 +5,10 @@ Namespace: SlottyMedia.Database.Repository.UserRepo
 Interface for the User Repository.
 
 ```csharp
-public interface IUserRepository : SlottyMedia.Database.IDatabaseRepository`1[[SlottyMedia.Database.Daos.UserDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
+public interface IUserRepository : SlottyMedia.Database.Repository.IDatabaseRepository`1[[SlottyMedia.Database.Daos.UserDao, SlottyMedia.Database, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]
 ```
 
-Implements [IDatabaseRepository&lt;UserDao&gt;](./slottymedia.database.idatabaserepository-1.md)
+Implements [IDatabaseRepository&lt;UserDao&gt;](./slottymedia.database.repository.idatabaserepository-1.md)
 
 ## Methods
