+++ slottymedia-docs/docs/code/SlottyMedia.DatabaseSeeding/index.md
@@ -2,6 +2,8 @@
 
 ## SlottyMedia.DatabaseSeeding
 
+[ImageDownloader](./slottymedia.databaseseeding.imagedownloader.md)
+
 [Login](./slottymedia.databaseseeding.login.md)
 
 [Rules](./slottymedia.databaseseeding.rules.md)
